create new model' Bookstore'
