+++ src/main/java/com/xebia/services/Bookstore/BookstoreService.java
@@ -1,0 +1,3 @@
+package com.xebia.services.Bookstore;
+
+public interface BookstoreService {}
