+++ src/main/java/com/xebia/services/Bookstore/BookstoreServiceImp.java
@@ -1,0 +1,3 @@
+package com.xebia.services.Bookstore;
+
+public class BookstoreServiceImp {}
